ecus/autopilot/python_requirements.txt
numpy
requests
tornado
websocket-client
carla==0.9.13
